requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
requests
python-dotenv
streamlit
pydantic
flask
flask-cors
langchain-openai
langchain